Change User Role mutation moved to Dialog
